# swagger.json
{
  "openapi": "3.0.0",
  "components": {
    "securitySchemes": {
      "bearerAuth" : {  
        "type" : "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security":{
    "bearerAuth" : []
  },
  "info": {
    "title": "Proyecto Flutter",
    "description": "API para registrar, listar, y editar Barberos, Clientes, Usuarios y Servicios",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": " http://127.0.0.1:4000"
    },
    {
      "url": "http://10.2.125.234:4000"
    }
  ],
...
  "paths": {
...
    "/obtenerServicios": {
      "get": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags":["Servicio"],
        "summary": "Listar todos los Usuarios",
        "description":"Este endpoint permite obtener un listado de los servicios registrados en el sistema",
        "responses": {
          "200": {
            "description": "Respuesta exitosa que contiene la lista de servicios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "serv_id"        : { "type": "integer" },
                      "serv_tipo"      : { "type": "string" },
                      "serv_precio"    : { "type": "string" }
                    },
                    "example": {
                      "serv_id"     : 2,
                      "serv_tipo"   : "solo barba",
                      "serv_precio" : "12000"
                    }                    
                  }
                }
              }
            }
          },
          "404": {
            "description": "error de servicio inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "mensaje": { "type": "string" }
                    },
                    "example": {
                      "mensaje": "Ningun servicio obtenido."
                    }            
                  }
                }
              }
            }
          }
        }
      }
    },
    "/registrarServicio": {
      "post": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags":["Servicio"],
        "summary": "Registrar un nuevo servicio",
        "description": "Este endpoint permite registrar un nuevo servicio. Requiere enviar un objeto JSON con los campos serv_tipo y serv_precio.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["'serv_tipo', 'serv_precio'"],
                "properties": {
                  "serv_tipo"      : { "type": "string" },
                  "serv_precio"    : { "type": "string" }
                },
                "example": {
                  "serv_tipo"   : "solo barba",
                  "serv_precio" : "12000"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": " Servicio registrado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": " Se ha registrado el Servicio"
                  }
                }
              }
            }
          },
          "400": {
            "description": "error de campos faltantes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Faltan campos por rellenar"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/editarServicio/{serv_id}": {
      "put": {
        "security":[{
          "bearerAuth" : []
        }],
        "tags":["Servicio"],
        "summary": "Editar un usuario existente",
        "description": "Este endpoint permite editar un servicio existente mediante su ID. Requiere un objeto JSON con los campos serv_tipo y serv_precio.",
        "parameters": [
          {
            "name": "serv_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["'serv_tipo', 'serv_precio'"],
                "properties": {
                  "serv_tipo"      : { "type": "string" },
                  "serv_precio"    : { "type": "string" }
                },
                "example": {
                  "serv_tipo"   : "solo barba",
                  "serv_precio" : "12000"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Devuelve un mensaje indicando que el Servicio editado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "se ha editado el servicio"
                  }
                }
              }
            } 
          },
          "400": {
            "description": "error de campos faltantes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" }
                  },
                  "example": {
                    "mensaje": "Faltan campos por rellenar"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}